Renamed sign-in-sign-up-form.jsx to auth-form.jsx
